Update 6
Fixed input function so it correctly updates time variables.  Added break input fields in html.  Added logic to check if break timer is on and if so to update the time variables with the break time inputs.  Added functionality for when the timers reach zero in regular or break mode.  Mostly working.